Move Complex toolkit to a separate package
